requirements.txt
unstructured[pdf]
langchain
langchain_community
langchain-google-genai
langchain_huggingface
langchain-qdrant
qdrant-client
google-genai
tiktoken
pymongo
python-dotenv
fastapi[all]
pydantic